msm: bam_dmux: tune downlink polling parameters
This change disables the BAM DMUX adaptive timer - this is to address
concerns about drops in TCP throughput for some use cases with the timer
enabled.  It also sets the aggregation to 3ms to save power while not
compromising performance. Power tests show a savings of 3-4% (compared to
having the timer poll the BAM pipe every 1 ms) for LTE (40Mbps downlink and
25Mbps uplink) and HSPA (7Mbps downlink and 384Kbps uplink) with no
degradation in performance.

Signed-off-by: Anurag Singh <anursing@codeaurora.org>
Change-Id: Ia51eae773723e60d8b051766ad80bfb0a16f5f6d